refactor(anytls-inbound): split mod.rs into focused modules
959-line mod.rs → 5 focused files:
- framing.rs: frame codec (read_frame/write_frame) + CMD_* constants
- tls.rs:     TLS acceptor construction + ephemeral rcgen cert
- user.rs:    SHA-256 password hash map builder
- handler.rs: handle_connection/fallback/udp_session/tcp_relay
- mod.rs:     thin — structs, new(), Drop, InboundHandlerTrait

No behavior changes. All 16 unit tests + smoke test pass.
